Update MakerClowdSquirrel.ts
for the "framework", net45 through net48, and vcredist are all static. for netcore (so netcoreapp3.1 through net7.0) this is dynamic. you can provide any value, eg. net6.0.0, net6.0.6, net 5.0.10, net7.0 even as soon as it comes out of preview. all of these default to x64 by default, but I think if you're looking for a pattern for typings that makes the most sense, we should just add the most common (both x86 and x64 variation).